# server/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Check"]}},"/gateways":{"post":{"operationId":"GatewaysController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGatewayDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gateway"}}}}},"tags":["Gateways"]},"get":{"operationId":"GatewaysController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Gateway"}}}}}},"tags":["Gateways"]}},"/gateways/{id}/add":{"put":{"operationId":"GatewaysController_addDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gateway"}}}}},"tags":["Gateways"]}},"/gateways/{id}/remove":{"put":{"operationId":"GatewaysController_removeDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gateway"}}}}},"tags":["Gateways"]}},"/gateways/{id}":{"get":{"operationId":"GatewaysController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gateway"}}}}},"tags":["Gateways"]},"delete":{"operationId":"GatewaysController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gateway"}}}}},"tags":["Gateways"]}},"/devices":{"post":{"operationId":"DevicesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeviceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}},"tags":["Devices"]},"get":{"operationId":"DevicesController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Device"}}}}}},"tags":["Devices"]}},"/devices/{id}":{"get":{"operationId":"DevicesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}},"tags":["Devices"]},"delete":{"operationId":"DevicesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}}}},"tags":["Devices"]}}},"info":{"title":"Device management","description":"You can download interfaces via the below link","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateGatewayDto":{"type":"object","properties":{"sn":{"type":"string","example":"sn-123456","description":"Serial Number"},"name":{"type":"string","example":"Gateway-1","description":"Name"},"ip4":{"type":"string","example":"192.168.10.12","description":"IP address"},"devices":{"example":[],"description":"Devices","type":"array","items":{"type":"string"}}},"required":["sn","name","ip4","devices"]},"Gateway":{"type":"object","properties":{"sn":{"type":"string","example":"sn-123456","description":"Serial Number"},"name":{"type":"string","example":"Some gateway","description":"Human readable name"},"ip4":{"type":"string","example":"8.8.8.8","description":"IP-address"},"devices":{"example":"[1111, 2222, 3333]","description":"IP-address","type":"array","items":{"type":"string"}}},"required":["sn","name","ip4","devices"]},"CreateDeviceDto":{"type":"object","properties":{"uid":{"type":"number","example":1234567,"description":"Uid"},"vendor":{"type":"string","example":"TP-Link","description":"Vendor"},"date":{"type":"number","example":1688193506,"description":"Time (timestamp)"},"status":{"type":"boolean","example":true,"description":"Status (online/offline)"}},"required":["uid","vendor","date","status"]},"Device":{"type":"object","properties":{"uid":{"type":"number","example":123456,"description":"Uid"},"vendor":{"type":"string","example":"TP link","description":"vendor"},"date":{"type":"number","example":1688174565,"description":"Date"},"status":{"type":"boolean","example":true,"description":"status"}},"required":["uid","vendor","date","status"]}}},"externalDocs":{"description":"swagger-spec.json","url":"http://localhost:8080/api-json"}}